requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.3.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==6.1.0
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
cliapp==1.20180812.1
click==8.1.7
cmdtest==0.32+git
colorama==0.4.6
comm==0.2.2
command-not-found==0.3
crypto==1.4.1
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.14
deptry==0.16.1
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
docopt==0.6.2
docutils==0.17.1
drf-yasg==1.21.7
duplicity==0.8.21
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.0
fabric==2.6.0
Fabric3==1.14.post1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-standalone-docs==0.1.6
fasteners==0.14.1
fastjsonschema==2.19.1
feedparser==6.0.8
filelock==3.13.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Babel==2.0.0
Flask-Cors==3.0.8
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.29.1
fqdn==1.5.1
frozenlist==1.2.0
fs==2.4.12
future==0.18.2
gast==0.5.2
git-filter-repo==2.45.0
gitsome==0.8.0
greenlet==3.0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
imagesize==1.3.0
importlib-metadata==4.6.4
inflection==0.5.1
invoke==1.7.3
ipykernel==6.29.4
ipython==7.31.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.3
joblib==1.3.2
jose==1.0.0
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
motor==3.4.0
mpmath==1.3.0
multidict==5.1.0
mypy==0.942
mypy-extensions==0.4.3
mysql-connector-python==8.2.0
mysqlclient==2.1.1
Naked==0.1.32
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
netifaces==0.11.0
nltk==3.8.1
notebook_shim==0.2.4
numpy==1.24.0
numpydoc==1.2
oauthlib==3.2.0
olefile==0.46
orjson==3.10.3
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parameterized==0.9.0
paramiko==2.9.3
parso==0.8.1
passlib==1.7.4
pathlib2==2.3.7.post1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==3.11.0
ply==3.11
prometheus_client==0.20.0
prompt-toolkit==3.0.28
protobuf==4.21.12
proton-core==0.1.16
proton-keyring-linux==0.0.2
proton-keyring-linux-secretservice==0.0.2
proton-vpn-api-core==0.27.2
proton-vpn-connection==0.14.4
proton-vpn-gtk-app==4.3.3
proton-vpn-killswitch==0.4.0
proton-vpn-killswitch-network-manager==0.4.5
proton-vpn-logger==0.2.1
proton-vpn-network-manager==0.4.2
proton-vpn-network-manager-openvpn==0.0.7
psutil==5.9.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.0
pycairo==1.20.1
pycodestyle==2.6.0
pycparser==2.21
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
pymongo==4.7.2
PyMySQL==1.1.0
PyNaCl==1.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.20.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.0.1
python-gnupg==0.4.8
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.9
pythran==0.10.0
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.3
referencing==0.32.1
regex==2023.12.25
reportlab==3.6.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
roman==3.3
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.4.2
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.2
sentry-sdk==1.4.3
setproctitle==1.2.2
sgmllib3k==1.0.0
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.3.1
Sphinx==4.3.2
SQLAlchemy==2.0.23
sqlparse==0.5.0
ssh-import-id==5.11
starlette==0.37.2
stripe==10.1.0
sympy==1.12
systemd-python==234
tabulate==0.8.9
terminado==0.18.1
textblob==0.18.0.post0
threadpoolctl==3.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
ttystatus==0.38
typed-ast==1.4.3
typer==0.12.3
types-aiofiles==0.1
types-annoy==1.17
types-appdirs==1.4
types-atomicwrites==1.4
types-aws-xray-sdk==2.8
types-babel==2.9
types-backports-abc==0.5
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.10
types-bleach==4.1
types-boto==2.49
types-braintree==4.11
types-cachetools==4.2
types-caldav==0.8
types-certifi==2020.4
types-characteristic==14.3
types-chardet==4.0
types-click==7.1
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-contextvars==0.1
types-croniter==1.0
types-cryptography==3.3
types-dataclasses==0.1
types-dateparser==1.0
types-DateTimeRange==0.1
types-decorator==0.1
types-Deprecated==1.2
types-docopt==0.6
types-docutils==0.17
types-editdistance==0.5
types-emoji==1.2
types-entrypoints==0.3
types-enum34==1.1
types-filelock==3.2
types-first==2.0
types-Flask==1.1
types-freezegun==1.1
types-frozendict==0.1
types-futures==3.3
types-html5lib==1.1
types-httplib2==0.19
types-humanfriendly==9.2
types-ipaddress==1.0
types-itsdangerous==1.1
types-JACK-Client==0.1
types-Jinja2==2.11
types-jmespath==0.10
types-jsonschema==3.2
types-Markdown==3.3
types-MarkupSafe==1.1
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.0
types-oauthlib==3.1
types-orjson==3.6
types-paramiko==2.7
types-Pillow==8.3
types-polib==1.1
types-prettytable==2.1
types-protobuf==3.17
types-psutil==5.8
types-psycopg2==2.9
types-pyaudio==0.2
types-pycurl==0.1
types-pyfarmhash==0.2
types-Pygments==2.9
types-PyMySQL==1.0
types-pyOpenSSL==20.0
types-pyRFC3339==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-dateutil==2.9.0.20240316
types-python-gflags==3.1
types-python-nmap==0.6
types-python-slugify==5.0
types-pytz==2021.1
types-pyvmomi==7.0
types-PyYAML==5.4
types-redis==3.5
types-requests==2.25
types-retry==0.9
types-selenium==3.141
types-Send2Trash==1.8
types-setuptools==57.4
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-stripe==2.59
types-tabulate==0.8
types-termcolor==1.1
types-toml==0.10
types-toposort==1.6
types-ttkthemes==3.2
types-typed-ast==1.4
types-tzlocal==0.1
types-ujson==0.1
types-vobject==0.9
types-waitress==0.1
types-Werkzeug==1.0
types-xxhash==2.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.10.0
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
usb-creator==0.3.7
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.24.6
wadllib==1.3.6
wakatime==14.0.1
watchfiles==0.21.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wordcloud==1.9.3
wrapt==1.16.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
xonsh==0.11.0
yarl==1.7.2
zipp==1.0.0
